+++ R/dsemRTMB.R
@@ -74,7 +74,7 @@ function( sem,
   # General warnings
   if( isFALSE(control$quiet) ){
     tsdata_SD = apply( tsdata, MARGIN=2, FUN=sd, na.rm=TRUE )
-    if( any((max(tsdata_SD,rm.na=TRUE)/min(tsdata_SD,rm.na=TRUE)) > 100) ){
+    if( isTRUE( (max(tsdata_SD,na.rm=TRUE)/min(tsdata_SD,na.rm=TRUE)) > 10) ){
       warning("Some variables in `tsdata` have much higher variance than others. Please consider rescaling variables to prevent issues with numerical convergence.")
     }
   }
@@ -145,8 +145,10 @@ function( sem,
   # Construct map
   if( is.null(control$map) ){
     Map = list()
+    # Map off x_tj for fixed when data is available
     Map$x_tj = factor(ifelse( is.na(as.vector(y_tj)) | (family[col(y_tj)] %in% c("normal","binomial","poisson","gamma")), seq_len(n_k), NA ))
-    Map$lnsigma_j = factor( ifelse(family=="fixed", NA, seq_along(Params$lnsigma_j)) )
+    # Map off sigma_j for fixed / bernoulli / Poisson
+    Map$lnsigma_j = factor( ifelse(family %in% c("fixed","binomial","poisson"), NA, seq_along(Params$lnsigma_j)) )
 
     # Map off mean for latent variables
     Map$mu_j = factor( ifelse(colSums(!is.na(y_tj))==0, NA, seq_len(n_j)) )
@@ -198,14 +200,9 @@ function( sem,
 
   # Export stuff
   if( control$run_model==FALSE ){
+    class(out) = "dsem"
     return( out )
   }
-
-  # Fit
-  #out$opt = fit_tmb( obj,
-  #                   quiet = control$quiet,
-  #                   control = list(eval.max=10000, iter.max=10000, trace=ifelse(control$quiet==TRUE,0,1) ),
-  #                   ... )
 
   # Optimize
   out$opt = list( "par"=obj$par )