+++ R/utility.R
@@ -126,3 +126,57 @@ function( object,
   if(what=="samples") return( resid_tjr )
   if(what=="loo") return( df )
 }
+
+#' @title Calculate total effects
+#'
+#' @description Calculate a data frame of total effects, representing the 
+#' estimated effect of every variable on every other variable and any time-lag
+#' from 0 (simultaneous effects) to a user-specified maximum lag.
+#'
+#' @param object Output from \code{\link{dsem}}
+#' @param n_lags Number of lags over which to calculate total effects
+#'
+#' @details
+#' Total effects are taken from the Leontief matrix \eqn{\mathbf{(I-P)^{-1}}},
+#' where \eqn{\mathbf{P}} is the path matrix across variables and times.
+#'
+#' @return
+#' A data frame listing the time-lag (lag), variable that is undergoing some 
+#' exogenous change (from), and the variable being impacted (to), along with the 
+#' total effect (total_effect) including direct and indirect pathways, and the
+#' partial "direct" effect (direct_effect)
+#'
+#' @export
+total_effect <-
+function( object,
+          n_lags = 4 ){
+
+  # Extract path matrix
+  Z = object$internal$tsdata
+  P_kk = make_matrices( 
+    beta_p = object$internal$parhat$beta,
+    model = object$sem_full,
+    times = seq_len(n_lags),
+    variables = colnames(Z)
+  )$P_kk            
+
+  # Define innovations
+  delta_kj = kronecker( Diagonal(n=ncol(Z)), 
+                        sparseMatrix(i=1, j=1, x=1, dims=c(n_lags,1)) )
+  IminusRho_kk = Diagonal(n=nrow(P_kk)) - P_kk
+  
+  # Calculate partial effect
+  Partial_kj = P_kk %*% delta_kj
+
+  # Calculate total effect using sparse matrices
+  Total_kj = solve( IminusRho_kk, delta_kj )
+  
+  # Make into data frame
+  out = expand.grid( "lag" = seq_len(n_lags)-1, 
+                     "to" = colnames(Z), 
+                     "from" = colnames(Z) )
+  out$total_effect = as.vector(Total_kj)
+  out$direct_effect = as.vector(Partial_kj)
+  return(out)
+}
+